refactor(worker): split live session binding
